Move custom decorator tests to separate directory
