+++ test/unit/decorators/customConstructor.spec.ts
@@ -1,5 +1,5 @@
-import * as TSTLErrors from "../../src/TSTLErrors";
-import * as util from "../util";
+import * as TSTLErrors from "../../../src/TSTLErrors";
+import * as util from "../../util";
 
 test("CustomCreate", () => {
     const luaHeader = `
